add modify and delete room methods

+++ LibrarySystem-Backend/src/main/java/ca/mcgill/ecse321/librarysystem/service/RoomService.java
@@ -24,9 +24,26 @@ public class RoomService {
 	// creates room, returns it so we know it's not null 
 	@Transactional 
 	public Room createRoom(
+			String currentUserId,
 			String roomNum,
 		    Integer capacity) 
 	{
+		// check permission: only head librarian have permission to modify the rooms
+		HeadLibrarian currentHeadLibrarian = headLibrarianRepository.findUserByIdNum(currentUserId);
+	    if (
+	      (currentHeadLibrarian == null || !currentHeadLibrarian.getIsLoggedIn())
+	    ) {
+	      throw new IllegalArgumentException(
+	        "You do not have permission to create a room"
+	      );
+	    }
+	    
+	    // check if newRoomNum is available
+	    for (Room room: toList(roomRepository.findAll()) )  {
+	    	if ( room.getRoomNum().equalsIgnoreCase(roomNum) ) throw new IllegalArgumentException("The room number already exists");
+	    }
+	    
+	    
 		Room room = new Room();
 	    room.setRoomNum(roomNum);
 	    room.setCapacity(capacity);
@@ -49,32 +66,49 @@ public class RoomService {
 	}
 
 	
-	// TODO add method for modify rooms
-	// only let head librarian modify rooms
+	// method for modify rooms
+	// only head librarian have permission to modify the rooms, they can modify the room number and the capacity, but the room numbers must be unique
+	@Transactional 
+	public Room updateRoom (String currentUserId, String oldRoomNum, String newRoomNum, int newCapacity) 
+	{
+		// check permission: only head librarian have permission to modify the rooms
+		HeadLibrarian currentHeadLibrarian = headLibrarianRepository.findUserByIdNum(currentUserId);
+	    if (
+	      (currentHeadLibrarian == null || !currentHeadLibrarian.getIsLoggedIn())
+	    ) {
+	      throw new IllegalArgumentException(
+	        "You do not have permission to update a room"
+	      );
+	    }
+		
+	    // check if newRoomNum is available
+	    for (Room room: toList(roomRepository.findAll()) )  {
+	    	if ( room.getRoomNum().equalsIgnoreCase(newRoomNum) ) throw new IllegalArgumentException("The new room number already exists");
+	    }
+		Room toUpdate = roomRepository.findRoomByRoomNum(oldRoomNum);
+		toUpdate.setCapacity(newCapacity);
+		toUpdate.setRoomNum(newRoomNum);
+		
+
+	    return toUpdate;		
+	}
 	
-	// TODO add method for delete rooms
+	// method for delete rooms
 	@Transactional 
 	public Room deleteRoom (String currentUserId, String roomNum) 
 	{
-		Librarian currentLibrarian = librarianRepository.findUserByIdNum(
-			      currentUserId
-			    );
-			    HeadLibrarian currentHeadLibrarian = headLibrarianRepository.findUserByIdNum(
-			      currentUserId
-			    );
-			    if (
-			      currentLibrarian == null ||
-			      !currentLibrarian.getIsLoggedIn() &&
-			      (currentHeadLibrarian == null || !currentHeadLibrarian.getIsLoggedIn())
-			    ) {
-			      throw new IllegalArgumentException(
-			        "You do not have permission to create an item reservation"
-			      );
-			    }
+		HeadLibrarian currentHeadLibrarian = headLibrarianRepository.findUserByIdNum(currentUserId);
+	    if (
+	      (currentHeadLibrarian == null || !currentHeadLibrarian.getIsLoggedIn())
+	    ) {
+	      throw new IllegalArgumentException(
+	        "You do not have permission to delete a room"
+	      );
+	    }
 			    
 		Room toDelete = roomRepository.findRoomByRoomNum(roomNum);
+		if (toDelete == null) throw new IllegalArgumentException("The room number is invalid");
 		roomRepository.delete(toDelete);
-
 	    return toDelete;		
 	}
